Extract documentation from Component methods
This adds documentation for methods of react components. It works with both React.createClass and classes. It reads from jsdoc blocks for descriptions and flow types for type information.

It supports:

- name
- description
- docblock (the raw docblock for further parsing)
- parameters (name, type, description)
- modifiers ('static', 'generator', 'async')
- return (type, description)

It only removes react component methods and leaves removing private methods to the users if they choose so.

Tested on the react-native codebase.

Closes #17